+++ src/controller/line-management/add-line.js
@@ -1,0 +1,52 @@
+import { rerenderSectionButtons } from '../section-management/index.js';
+import { addLineLocalStorage, addLineList } from './model.js';
+import { checkValidLineSelector } from './check.js';
+import { removeLine } from './remove-line.js';
+import { storage } from '../../model/index.js';
+import { addButtonEvent } from '../utils.js';
+import { elementIds } from '../../utils.js';
+import { initLineInput } from './init.js';
+import { renderLine } from './view.js';
+
+const addLineButtonEvent = (lineName) => {
+  for (let item of document.querySelectorAll(
+    `table[id=line-table] tbody tr td button`,
+  )) {
+    if (item.dataset.line === `${lineName}-button`) {
+      addButtonEvent(item, removeLine);
+    }
+  }
+};
+
+export const addLine = () => {
+  const lineName = elementIds.lineNameInput.value;
+  const lineStartStationSelectorValue =
+    elementIds.lineStartStationSelector.value;
+  const lineEndStationSelectorValue = elementIds.lineEndStationSelector.value;
+  if (
+    !checkValidLineSelector(
+      lineStartStationSelectorValue,
+      lineEndStationSelectorValue,
+    )
+  ) {
+    return;
+  }
+  if (
+    !addLineLocalStorage(
+      lineStartStationSelectorValue,
+      lineEndStationSelectorValue,
+    )
+  ) {
+    return;
+  }
+  renderLine({
+    lineName,
+    lineStartStationSelectorValue,
+    lineEndStationSelectorValue,
+  });
+  initLineInput();
+  addLineList(lineStartStationSelectorValue, lineEndStationSelectorValue);
+  addLineButtonEvent(lineName);
+  rerenderSectionButtons();
+  console.log(storage.getLocalStorageMap('subway-line'));
+};
