+++ src/controller/line-management/init.js
@@ -1,0 +1,6 @@
+import { elementIds } from '../../utils.js';
+
+export const initLineInput = () => {
+  elementIds.lineNameInput.value = '';
+  elementIds.lineNameInput.focus();
+};
